issue #2870: fix some warning

+++ cocos/2d/CCLabelBMFont.cpp
@@ -936,7 +936,7 @@ void LabelBMFont::updateLabel()
             Sprite* characterSprite;
             long justSkipped = 0;
             
-            while (!(characterSprite = static_cast<Sprite*>( this->getChildByTag(j + skip + justSkipped))) )
+            while (!(characterSprite = static_cast<Sprite*>( this->getChildByTag(static_cast<int>(j + skip + justSkipped)))) )
             {
                 justSkipped++;
             }
