+++ cocos/editor-support/spine/Atlas.cpp
@@ -136,7 +136,7 @@ static int readTuple (const char* end, Str tuple[]) {
 }
 
 static char* mallocString (Str* str) {
-	int length = str->end - str->begin;
+	long length = str->end - str->begin;
 	char* string = MALLOC(char, length + 1);
 	memcpy(string, str->begin, length);
 	string[length] = '\0';
@@ -144,7 +144,7 @@ static char* mallocString (Str* str) {
 }
 
 static int indexOf (const char** array, int count, Str* str) {
-	int length = str->end - str->begin;
+	long length = str->end - str->begin;
 	int i;
 	for (i = count - 1; i >= 0; i--)
 		if (strncmp(array[i], str->begin, length) == 0) return i;
@@ -156,7 +156,7 @@ static int equals (Str* str, const char* other) {
 }
 
 static int toInt (Str* str) {
-	return strtol(str->begin, (char**)&str->end, 10);
+	return static_cast<int>(strtol(str->begin, (char**)&str->end, 10));
 }
 
 static Atlas* abortAtlas (Atlas* self) {
@@ -168,10 +168,10 @@ static const char* formatNames[] = {"Alpha", "Intensity", "LuminanceAlpha", "RGB
 static const char* textureFilterNames[] = {"Nearest", "Linear", "MipMap", "MipMapNearestNearest", "MipMapLinearNearest",
 		"MipMapNearestLinear", "MipMapLinearLinear"};
 
-Atlas* Atlas_readAtlas (const char* begin, int length, const char* dir) {
-	int count;
+Atlas* Atlas_readAtlas (const char* begin, long length, const char* dir) {
+	long count;
 	const char* end = begin + length;
-	int dirLength = strlen(dir);
+	long dirLength = static_cast<long>(strlen(dir));
 	int needsSlash = dirLength > 0 && dir[dirLength - 1] != '/' && dir[dirLength - 1] != '\\';
 
 	Atlas* self = NEW(Atlas);
@@ -283,9 +283,9 @@ Atlas* Atlas_readAtlas (const char* begin, int length, const char* dir) {
 }
 
 Atlas* Atlas_readAtlasFile (const char* path) {
-	int dirLength;
+	long dirLength;
 	char *dir;
-	int length;
+	long length;
 	const char* data;
 
 	Atlas* atlas = 0;
